refactor(tests): separate test files, debug test failure
